refactor: ProjectSideBar를 SideBarTemplate 컴포넌트 내부로 이동
