--- frontend/manage/src/components/pages/SideBarTemplate/styles.tsx
@@ -1,12 +1,0 @@
-import styled from "styled-components";
-import { contentBoxCSS } from "../../../styles/css";
-
-export const Container = styled.div`
-  display: flex;
-  justify-content: space-between;
-`;
-
-export const MainContent = styled.main`
-  width: 50rem;
-  ${contentBoxCSS}
-`;
